--- src/backend/efiling-worker/src/main/java/ca/bc/gov/open/jag/efilingworker/service/CSOService.java
@@ -1,8 +1,0 @@
-package ca.bc.gov.open.jag.efilingworker.service;
-
-import ca.bc.gov.open.jag.ag.csows.filing.FilingPackage;
-import ca.bc.gov.open.jag.ag.csows.filing.SubmitFilingResponse;
-
-public interface CSOService {
-   SubmitFilingResponse submitFiling(FilingPackage filingPackage);
-}
