+++ src/backend/libs/efiling-facade-client/pom.xml
@@ -11,7 +11,7 @@
         <relativePath/>
     </parent>
 
-    <groupId>ca.bc.gov.open</groupId>
+    <groupId>ca.bc.gov.open.jag</groupId>
     <artifactId>efiling-facade-client</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>efiling-facade-client</name>
