+++ src/backend/libs/efiling-facade-client/src/main/java/ca/bc/gov/open/jag/efilingfacadeclient/EfilingFacadeService.java
@@ -1,0 +1,11 @@
+package ca.bc.gov.open.jag.efilingfacadeclient;
+
+import ca.bc.gov.open.jag.ag.csows.filing.FilingPackage;
+import ca.bc.gov.open.jag.ag.csows.filing.SubmitFilingResponse;
+
+import java.math.BigDecimal;
+
+public interface EfilingFacadeService {
+
+    public SubmitFilingResponse submitFiling(FilingPackage filingPackage);
+}
