+++ src/backend/libs/efiling-facade-client/src/main/java/ca/bc/gov/open/jag/efilingfacadeclient/MockCSOEfilingServiceImpl.java
@@ -1,14 +1,15 @@
-package ca.bc.gov.open.jag.efilingworker.service;
+package ca.bc.gov.open.jag.efilingfacadeclient;
 
 import ca.bc.gov.open.jag.ag.csows.filing.FilingPackage;
 import ca.bc.gov.open.jag.ag.csows.filing.SubmitFilingResponse;
 
 import java.math.BigDecimal;
 
-public class MockCSOService implements CSOService {
+public class MockCSOEfilingServiceImpl implements EfilingFacadeService {
 
     @Override
     public SubmitFilingResponse submitFiling(FilingPackage filingPackage) {
+
         SubmitFilingResponse submitFilingResponse = new SubmitFilingResponse();
         submitFilingResponse.setReturn(new BigDecimal(1));
         return submitFilingResponse;
